Move the response objects to functional tests module
